% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mic.R
\name{MIC}
\alias{MIC}
\title{Maximal information coefficient (MIC)}
\usage{
MIC(x, y, R = 100, ...)
}
\arguments{
\item{x}{A vector, matrix or numeric data frame}

\item{y}{A vector, matrix or numeric data frame}

\item{R}{Number of replications that will be used to calculate the p-value}

\item{...}{Additional parameters defined in the code{mine} function of the
code{\link{minerva}} package.}
}
\description{
This function calculates the Maximal information coefficient (MIC).
Calculate the dependence of pairs of variables. The MIC belongs to a larger
classification of nonparametric exploration statistics based on maximum
information (MINE).
}
\details{
It also provides the p-value for the independence hypothesis, that are
generated by permuting the data and seeing how likely it is that the observed
MIC value arises from the perturbed data.
}
\examples{
#Lineal dependence
x<-rnorm(200, sd=2.5)
y<-2*x+rnorm(200, sd=2.5)
MIC(x, y, R=100)

#No dependency
x<-rnorm(200, mean=2, sd =1.8)
y<-rnorm(200, mean=0, sd =2.5)
MIC(x, y, R=100)
}
