% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdc.R
\name{rdc}
\alias{rdc}
\title{Randomized Dependence Coefficient (RDC)}
\usage{
rdc(x, y, k = 20, s = 1/6, f = sin)
}
\arguments{
\item{x}{A vector, matrix or numeric data frame}

\item{y}{A vector, matrix or numeric data frame}

\item{k}{Number of non-linear projections of the copula, by default k = 20}

\item{s}{Variance to draw i.i.d. projection coefficients in N ~(0, sI),
by defect is 1/6}

\item{f}{Function that is used for the generation of random non-linear
projections, if it is not indicated it uses the sinusoidal projections (sin)}
}
\description{
This function calculates the Random Dependence Coefficient (RDC).
Calculate the dependence between the random samples as the highest canonical
correlation between the k random non-linear projections of their copula
transformations.
}
\details{
It also provides the p-value for the independence hypothesis, assuming the
normality of the data through the Bartlett's approximation.
}
\examples{
#Lineal dependence
x<-rnorm(200, sd=2.5)
y<-2*x+rnorm(200, sd=2.5)
rdc(x,y)

#No dependency
x<-rnorm(200, mean=2, sd =1.8)
y<-rnorm(200, mean=0, sd =2.5)
rdc(x,y)
}
